supplise Feature :finish supplies feature

+++ pet/models.py
@@ -14,7 +14,7 @@ class Pet(models.Model):
     type=models.CharField(max_length=70,blank=False)
     name_pet = models.CharField(blank=True,max_length=70,null=True)
     description = models.TextField(blank=True,max_length=500,null=True)
-    price = models.IntegerField(default=0,validators=[MinValueValidator(0)])
+    price = models.FloatField(default=0,validators=[MinValueValidator(0)])
     published = models.DateTimeField(default=timezone.now, blank=True)
     adoption=models.BooleanField(default=False)
     picture = models.ImageField(upload_to='pets', blank=True)
